requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
feature-engine
joblib
sweetviz
plotly
streamlit
Pillow
